Refactor: search 블루프린트 나누기 - 전시, 작가, 미술관을 각각 나눠 등록함
